+++ database/src/models/transaction.rs
@@ -3,7 +3,7 @@ use crate::models::types::payload::Payload;
 
 pub struct Transaction {
     pub transaction_id: Hash,
-    pub subnetwork_id: i16,
+    pub subnetwork_id: i32,
     pub hash: Hash,
     pub mass: i32,
     pub payload: Payload,
